Enabled Windows 7 and Windows Vista installation
I have enabled the installation on Windows 7 and Windows Vista. Note that Service Pack 1 is required for Windows 7 and Service Pack 2 is required for Windows Vista. Currently, there are no issues after installing Python 3.9 on these operating systems. If there will be any problems in the future, I will try to solve them.